refactored button to separate component
