+++ app/path-framework/page/element/button/link-button.component.ts
@@ -1,4 +1,4 @@
-import {Button} from "./button";
+import {Button} from "./button.component";
 import {PathService} from "../../../service/path.service";
 import {IPathApp} from "../../../pathinterface";
 import {Component, Input, Output} from "@angular/core";
