+++ app/path-framework/path.ts
@@ -4,7 +4,7 @@ export * from './path-app.component';
 export * from './page/page';
 
 export * from './page/element/page-element';
-export * from './page/element/button/button';
+export * from './page/element/button/button.component';
 export * from './page/element/button/button-detail';
 export * from './page/element/button/back-button.component';
 export * from './page/element/button/link-button.component';
